bug fixes #1
