+++ src/main/java/org/example/coworking/factory/CoworkingFactory.java
@@ -1,0 +1,5 @@
+package org.example.coworking.factory;
+
+public interface CoworkingFactory<T> {
+    T create();
+}
