+++ src/main/java/org/example/coworking/model/enums/UserRole.java
@@ -1,4 +1,4 @@
-package org.example.coworking.enums;
+package org.example.coworking.model.enums;
 
 public enum UserRole {
     USER,
