+++ src/main/java/org/example/coworking/service/BookingService.java
@@ -1,8 +1,8 @@
-package org.example.coworking.services;
+package org.example.coworking.service;
 
-import org.example.coworking.enums.ResourceType;
-import org.example.coworking.models.Booking;
-import org.example.coworking.models.User;
+import org.example.coworking.model.enums.ResourceType;
+import org.example.coworking.model.Booking;
+import org.example.coworking.model.User;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
