+++ src/test/java/BookingServiceTest.java
@@ -1,7 +1,7 @@
-import org.example.coworking.enums.ResourceType;
-import org.example.coworking.models.Booking;
-import org.example.coworking.models.User;
-import org.example.coworking.services.BookingService;
+import org.example.coworking.model.enums.ResourceType;
+import org.example.coworking.model.Booking;
+import org.example.coworking.model.User;
+import org.example.coworking.service.BookingService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import java.time.LocalDateTime;
