+++ src/test/java/WorkplaceTest.java
@@ -1,4 +1,4 @@
-import org.example.coworking.models.Workplace;
+import org.example.coworking.model.Workplace;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
